pKVM: VMX: Share vCPU state for EXIT_REASON_IO_INSTRUCTION
Share the necessary vCPU state for the host to emulate the IO
instructions. For the npVM, sharing the exit instruction length for the
host to skip the IO instruction. No need to share the RAX as this is
already shared. For the pVM, sharing the RAX with the host, and no need
to share the exit instruction length as the host will not skip
instruction for the pVM.

Squash changelog:
  d389171cfc69 ("pKVM: x86: Fix switch-case structure in share_nonprotected_vcpu_state")
  09cc0f5b25ff ("pKVM: VMX: Fix RAX register leak and injection in port-IO")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>