pKVM: VMX: Handle EXIT_REASON_CPUID
As the pKVM hypervisor maintains the guest CPUID entries, the
EXIT_REASON_CPUID is handled alone inside the pKVM hypervisor.

The guest should have a way to detect it is running in a pVM or a npVM
so that it can enable the pVM's enlightenment, emulating the pVM's CPUID
leaf KVM_CPUID_SIGNATURE with the signature "PKVMPKVMPKVM".

Squash changelog:
  dd48d3f6958a ("pKVM: x86: Fix KVM CPUID signature lookup to use dynamic base")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>