pKVM: VMX: Setup the RSP and RIP fields for the pKVM hypervisor
Each pkvm_pcpu has its own stack memory which is defined as 16K. The top
16 bytes are reserved to store parameters as below for the pkvm
hypervisor to use after the deprivileged host vmexit to the root mode.
And the RSP starts from stack top - 16 is:
 ------------ Stack layout ----------
 stack_top:
 stack_resv + 8:     struct vcpu_vmx *vmx
 stack_resv + 0:     pointer to vcpu->arch.regs
 stack_resv:         VMCS.HOST_RSP for PCPU
                     .........
                     .........
 stack_bottom:

The RIP is set to pkvm_host_vmexit_entry which is the first function to
run after vmexit. It will save all the GPRs to vcpu->arch.regs and
handle vmexit reasons (implemented by the following patch). Then restore
the GPRs and vmresume to the deprivileged host.

Mark the VM-Exit entry as the end of the unwind stack because VMX loads
RIP and RSP directly, without creating a conventional caller frame.

[ Vineeth: Create vcpu_regs.h ]

Squash changelog:
  d8cfc3c55b49 ("pKVM: x86: Clear guest GPRs after VM-Exit to prevent speculative leaks")
  c1a901671fde ("pKVM: x86: Index host_vcpus and pcpus via consecutive number")

Signed-off-by: Vineeth Pillai (Google) <vineeth@bitbyteword.org>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>