pKVM: VMX: Implement vm_init operation
Import the vmx_vm_init() from KVM implementation to support the vm_init
operation in the pKVM hypervisor.

The CPU which can run pKVM doesn't have L1TF vulnerability, which is
guaranteed by the pKVM deprivilege code. Thus checking for the L1TF
mitigation is not necessary for the pKVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>