pKVM: VMX: Implement handle_exit_irqoff operation
Import vmx_handle_exit_irqoff from KVM to implement handle_exit_irqoff
operation for the pKVM hypervisor. Although the irq is always off in the
pKVM, still implement this to reuse the vmexit handlers in this
operation. The detailed handlers for the EXIT_REASON_EXTERNAL_INTERRUPT
and the EXIT_REASON_EXCEPTION_NMI will be added later.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>