ANDROID: pkvm: x86: trace: Add tracing of host hypercalls
Currently pKVM's vmexit tracing provides perf stats for individual
vmexit reasons only, so in particular, it doesn't distinguish between
different hypercalls (vmexits with VMCALL reason). Whereas in practice
the vast majority of host vmexits to pKVM are hypercalls. So it would be
very useful to provide separate stats for different types of hypercalls
(i.e. based on the hypercall number in the RAX register). So provide it,
in addition to the existing vmexit stats.

No such support for guest hypercalls yet.

Bug: 459523218
Test: run VMs and run "echo 1 >/sys/kernel/debug/pkvm/set_vmexit_trace;
cat /sys/kernel/debug/pkvm/vmexit_trace"
Upstream-Task: 402758258
Ignore-Checkpatch: open square bracket is at the beginning of a macro
definition, not after an array identifier.

Change-Id: If51a20e5d663db74b9aefb6a436a34f741ecc581
Signed-off-by: Dmytro Maluka <dmaluka@google.com>