pKVM: x86: trace: Add debugfs for the pKVM host
Add a debugfs root for the pKVM host at /sys/kernel/debug/pkvm/. It will
be used by the pKVM host to register its own debugfs files.

As once the pKVM hypervisor starts to run, there is no way to disable it
thus it is not necessary to remove this debugfs root.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>