pKVM: x86: Add vm_destroy PV interface
Add vm_destroy PV interface to destroy a guest VM. This interface
releases the pkvm_vm handle, and involves the vendor-specific operation
to destroy vendor resources. It also tears down donated memory pages and
unshares the host KVM structure.

The physical address and size of the torn-down memory pages are stored
in the pkvm_memcache structure, which will be returned for the host to
reclaim.

Squash changelog:
  77a48d010d50 ("pKVM: x86: Propagate error from pkvm_vm_destroy() to host")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>