pKVM: x86: trace: Support printing the guest VM trace
Extend the vmexit_trace_show() for the debug file to print the guest
VM's trace. If a debug file has a struct kvm as its private data, it
means catting this debug file will show the trace of a specific guest
VM. Otherwise it will show the host VM trace and the trace of all guest
VMs.

Squash changelog:
  50dedc858c42 ("pKVM: x86: Fix uninitialized return value in vmexit trace printing")
  d10bebd956f7 ("pKVM: x86: trace: Check size to prevent allocating zero bytes")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>