SQUASHME: pKVM: x86: Make pkvm_hyp_donate_host() panic on any error
pKVM hypervisor should only call pkvm_hyp_donate_host() for those pages
that are guaranteed to be owned by the hypervisor (usually it means
pages that were previously donated from the host to the hypervisor via
pkvm_host_donate_hyp()), so if pkvm_hyp_donate_host() is called with
incorrect arguments, it means a bug in pKVM.

In practice we never check the return value of pkvm_hyp_donate_host(),
assuming that it should succeed.

So let pkvm_hyp_donate_host() BUG_ON in case of any error, and make it
void, similarly to pkvm_host_unshare_hyp().

Fixes: 40ad026472af ("pKVM: x86: Add memory page donation between host and hypervisor")
Signed-off-by: Dmytro Maluka <dmaluka@google.com>