pKVM: x86: Add MMIO donation from hyp to host
Similar with the page donation, support donating the MMIO pages from
hypervisor to host. A locked version API is provided for the scenario
which the host mmu is already locked.

Donating the MMIO pages from host to hypervisor is not necessary as
currently there is no requirement for that.

Squash changelog:
  51fc55efab4c ("pKVM: x86: Add overflow validation for address ranges in MMU operations")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>