pKVM: x86: Sort the exception table exception
Sort the exception table section so that can use binary search to search
the exception table for exception fixup.

As the pKVM hypervisor's IDT will be programmed into VMCS before
deprivileging the CPU, one deprivileging is done and the CPU enters to
the root mode, the pKVM's exception handlers should be functional. So
before that, sort pKVM's exception table to make sure the exception
fixup working as expected.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>