ANDROID: pkvm: x86: Hypercalls for domain pagetable management
Implement hypercalls for mapping/unmapping of IOVA addresses.

This patch implements the iommu map and unmap hypercall stubs, and
leaves the actual functionality to a later patch.

Host iommu driver has a performance optimization to delay iotlb cache
flush on unmap. This breaks pkvm security guarantee because the pages
freed by unmap may be donated to pvms and could be used to store
sensitve data. Until a flush happens the iotlb cache may contain stale
entries pointing to these pages and enables the devide to read those
pages.

So flush iotlb cache unconditionally on unmap for now. We could revisit
this later to figure out how to securely delay the and batch flushing
for performance.

Bug: 391542119
Upstream-Task: 402758258
Test: Build and Boot

Change-Id: I5f4e29919c78f6dfa381e4e353ec166c775a3230
Signed-off-by: Vineeth Pillai <vineethrp@google.com>
Signed-off-by: Aashish Sharma <shraash@google.com>