SQUASHME: pkvm: x86: Fix worstcase calculation in __pkvm_pgtable_max_pages()
__pkvm_pgtable_max_pages() is used for calculating the worst-case number
of pages needed for constructing page tables for mapping a VA range
containing the given number of pages (for example, for identity-mapping
a memblock range in the host EPT). However, the calculation it does is
a bit too optimistic in the general case, as the mapped range may cross
the boundary of an address range that can be mapped by a single page
table at the given level (e.g. 2MB for 4K pages) even if the range size
is within this boundary (e.g. smaller than 2MB), so an extra page table
may be needed to map the whole range. So account for that as well.

Fixes: 4eb1437d5939 ("pKVM: x86: Create memory pool for pKVM mmu")
Signed-off-by: Dmytro Maluka <dmaluka@google.com>