ANDROID: pkvm: x86: Fix worstcase calculation in __pkvm_pgtable_max_pages()
__pkvm_pgtable_max_pages() is used for calculating the worst-case number
of pages needed for constructing page tables in order to contiguously
map the given number of pages (for example, for identity-mapping a
memblock range in the host EPT). However, the calculation it does is
a bit too optimistic in the general case, as the mapped range may cross
the boundary of an address range that can be mapped by a single page
table at the given level (e.g. 2MB for 4K pages) even if the range size
is within this boundary (e.g. smaller than 2MB), so an extra page table
may be needed to map the whole range. So account for that as well.

Bug: 391542119
Upstream-Task: 402758258
Test: Build and Boot

Change-Id: I9c84de9227ed3a88fe606233b8cbb469007e4eba
Signed-off-by: Dmytro Maluka <dmaluka@google.com>