ANDROID: pkvm: x86: Pin/unpin physical pages on DMA map/unmap
Donation of host owned page(s) that is mapped for DMA to a pVM allows
devices to access pVM memory through the DMA mapping. This is a breach
of security guarantee by pKVM and should not be allowed.

Pin the pages(by incrementing hyp_page refcount) that are being mapped
for DMA. Before mapping, we should also make sure that the memory range
is owned by the host. __pkvm_use_dma() api does both these for us.

Similarly, during unmap, unpin the pages so that the host is able to
donate pages if it needs to. Unpinning the pages in the unmap path is a
bit more complex as unmap hypercall receives the IOVA range and we need
to walk the page table to determine the physical pages that are mapped.
Its possible to unpin while walking the pagetable during the unmap, but
unpinning before flushing the caches(iotlb and devtlb) is a security
risk as a compromised host would be able to donate this page after
unpin and device would be able to access unpinned pages during the small
window between unpin and cache flush. So we do a two pass approach where
we unmap the pages and flush the caches in the first pass and then unpin
the unmapped physical range in the second pass.

Bug: 391542119
Upstream-Task: 402758258
Test: Build and Boot

Change-Id: Ieaaa72586ec6aa9260d837a03e82aabf17677b75
Signed-off-by: Aashish Sharma <shraash@google.com>
Signed-off-by: Vineeth Pillai <vineethrp@google.com>