ANDROID: pkvm: x86: Use hardware ignored bit 63 to represent mapped PTE
Unmap is done in two passes. In the first pass, we unpresent the PTE and
then flush the caches(iotlb and devtlb). Then we do another pass to
unpin all the physical pages in the range that was unmapped.

So we need a way to determine if the PTE is mapped(even if unpresented).
Instead of relying on the SL/FL address in the PTE, use an ignored bit
(bit 63) to denote that the PTE is mapped.

Bug: 391542119
Upstream-Task: 402758258
Test: Build and Boot

Change-Id: I69799e3b77bea722fb942375ac3cdc9f4d98dda3
Signed-off-by: Vineeth Pillai <vineethrp@google.com>