pKVM: x86: Add event injection PV interfaces
Add event injection PV interfaces for the host to inject irq, nmi and
exceptions via vendor-specific operations.

The PV interfaces for injecting irq/nmi are always accessible to the
host as the host emulates the irq/nmi injections for both npVM and pVM.
The exception for pVM is emulated by the pKVM, thus making this PV
interface always inaccessible to the host.

Squash changelog:
  14b06d159c37 ("SQUASHME: pKVM: x86: Disallow the host to inject software interrupts to a pVM")
  75505446677c ("SQUASHME: pKVM: x86: Validate inject vector number for pVM")
  4487a872c966 ("pKVM: x86: Fix TOCTOU and shared_vcpu accessing for event injection")
  227d30f03a3f ("pKVM: x86: Don't WARN_ON if irq/nmi injection not allowed")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>