SQUASHME: pKVM: x86: s/init_finalize/init/
init_finalize hypercall name is a slight misnomer - the hypercall
basically initializes pKVM hypervisor and not just the finalization
part. More over, we would be using the finalize terminology later when
we handle undoing the initialization and reprivileging cpus in the case
of initialization failure.

So rename the hypercall and related functions in pKVM and host.

Signed-off-by: Vineeth Pillai (Google) <vineeth@bitbyteword.org>